refactor(chore): Refactor project structure splitting up domains
Fix discover mode
