+++ tap_github/streams.py
@@ -1,7 +1,15 @@
 # Python imports
+import os
+import json
 # Third-Party imports
 import singer
+from singer import metadata
 # Project imports
+from exceptions import *
+from settings import KEY_PROPERTIES
+
+
+logger = singer.get_logger()
 
 
 def get_catalog():
@@ -15,11 +23,27 @@ def get_catalog():
             'stream': schema_name,
             'tap_stream_id': schema_name,
             'schema': schema,
-            'metadata' : metadata.to_list(mdata),
+            'metadata': metadata.to_list(mdata),
             'key_properties': KEY_PROPERTIES[schema_name],
         }
         streams.append(catalog_entry)
     return {'streams': streams}
+
+
+def load_schemas():
+    try:
+        schemas = {}
+        for filename in os.listdir(get_abs_path('schemas')):
+            if '.json' in filename:
+                logger.info(f'Processing file {filename}')
+                path = get_abs_path('schemas') + '/' + filename
+                file_raw = filename.replace('.json', '')
+                with open(path) as file:
+                    schemas[file_raw] = json.load(file)
+        return schemas
+    except (UnicodeDecodeError, KeyError) as e:
+        logger.error(f"Error processing filename {filename} in path {path} with file {file_raw}")
+        raise SchemaFileFormatException(e)
 
 
 def get_selected_streams(catalog):
@@ -46,3 +70,42 @@ def get_stream_from_catalog(stream_id, catalog):
         if stream['tap_stream_id'] == stream_id:
             return stream
     return None
+
+
+def get_abs_path(path):
+    return os.path.join(os.path.dirname(os.path.realpath(__file__)), path)
+
+
+def generate_pr_commit_schema(commit_schema):
+    pr_commit_schema = commit_schema.copy()
+    pr_commit_schema['properties']['pr_number'] = {
+        "type":  ["null", "integer"]
+    }
+    pr_commit_schema['properties']['pr_id'] = {
+        "type": ["null", "string"]
+    }
+    pr_commit_schema['properties']['id'] = {
+        "type": ["null", "string"]
+    }
+    return pr_commit_schema
+
+
+def populate_metadata(schema_name, schema):
+    mdata = metadata.new()
+    #mdata = metadata.write(mdata, (), 'forced-replication-method', KEY_PROPERTIES[schema_name])
+    mdata = metadata.write(mdata, (), 'table-key-properties', KEY_PROPERTIES[schema_name])
+    for field_name in schema['properties'].keys():
+        if field_name in KEY_PROPERTIES[schema_name]:
+            mdata = metadata.write(mdata, ('properties', field_name), 'inclusion', 'automatic')
+        else:
+            mdata = metadata.write(mdata, ('properties', field_name), 'inclusion', 'available')
+    return mdata
+
+
+def write_metadata(mdata, values, breadcrumb):
+    mdata.append(
+        {
+            'metadata': values,
+            'breadcrumb': breadcrumb
+        }
+    )
